move and rename CartDialog
